Basic IRM features implemented
A shared folder now has a UUID and an "owner" (identified by teh email).
The owner can go to the website and change the permission of the folder
(identified by UUID) from private->shared and vice versa, and other
users will have to obey.

This is flimsy now (can be hacked by editing the .xml), but we can take
this info and encrypt it as well.